use sveltekit and vite for packaging and building, fix stroybook structure, and change dir structure for lib packaging
Still need to have better typing and to put the sveltekit example in the routes.
