--- .storybook/main.js
@@ -1,9 +1,0 @@
-module.exports = {
-  stories: ["../stories/**/*.stories.js"],
-  addons: [
-    "@storybook/addon-actions",
-    "@storybook/addon-links",
-    "@storybook/addon-storysource",
-    "@storybook/addon-notes/register",
-  ],
-};
