+++ package/.gitignore
@@ -1,0 +1,4 @@
+node_modules
+dist
+storybook-static
+.svelte-kit
