+++ package/index.d.ts
@@ -1,0 +1,2 @@
+import DropZone from "./internal/components/Dropzone.svelte";
+export default DropZone;
