+++ package/index.js
@@ -1,0 +1,3 @@
+// This file exists for Stories
+import DropZone from "./internal/components/Dropzone.svelte";
+export default DropZone;
