--- src/index.js
@@ -1,5 +1,0 @@
-// This file exists for Stories
-// TODO remove this file and use index.svelte only
-import DropZone from "./components/Dropzone.svelte";
-
-export default DropZone;
