--- src/index.svelte
@@ -1,3 +1,0 @@
-import DropZone from "./components/Dropzone.svelte";
-
-export default DropZone;
