+++ src/lib/index.ts
@@ -1,0 +1,4 @@
+// This file exists for Stories
+import DropZone from "./internal/components/Dropzone.svelte";
+
+export default DropZone;
