+++ stories/1-dropzone.stories.js
@@ -1,17 +1,13 @@
-import { action } from "@storybook/addon-actions";
-
 import BasicDropZoneView from "./views/BasicDropzoneView.svelte";
 import DisabledDropzoneView from "./views/DisabledDropzoneView.svelte";
 import CustomSlotDropzoneView from "./views/CustomSlotDropzoneView.svelte";
 import NoClickDropZoneView from "./views/NoClickDropzoneView.svelte";
 import NoDragDropzoneView from "./views/NoDragDropzoneView.svelte";
 
-import intro from "./../README.md";
-
 export default {
   title: "Examples",
   parameters: {
-    notes: { Introduction: intro },
+    notes: { Introduction: __readme__, 'Other': 'other' },
   },
 };
