+++ stories/2-advanced-dropzone.stories.js
@@ -1,14 +1,10 @@
-import { action } from "@storybook/addon-actions";
-
 import WorkingCSVFileUploadView from "./views/WorkingCSVFileUploadView.svelte";
 import FullyFeaturedDropzoneView from "./views/FullyFeaturedDropzoneView.svelte";
-
-import intro from "./../README.md";
 
 export default {
   title: "Advanced Examples",
   parameters: {
-    notes: { Introduction: intro },
+    notes: { Introduction: __readme__ },
   },
 };
